Reworked the parsing of $(cmd), $(<file) , $<(xcmd) , $<(<xfile). Parsing of the sub command is now parsed in the javacc parser entirely so that no lexical assumptions are made about the string values of "cmd". This means that ANY command can be the body of $() and $<().
However to implement this, I removed expansion of subproc commands from inside quotes,
and from IO redirects.  This is a backwards incompatible change, although hopefully 
not a severe one.   In exchange for some convenience I can now guarentee the $(cmd) syntax is always accurate, which could not easily be done before.
To get $(cmd) inside quotes or in an IO redirect assign it to a variable.

x=$(cmd)
echo "$x"